+++ src/mockData/favorites.ts
@@ -1,0 +1,119 @@
+import { TOffer } from '../types';
+
+export const favOffersList: TOffer[] = [
+  {
+    id: 'ef79442e-d9a6-4e0f-8e72-92b70cdb892e',
+    title: 'Tile House',
+    type: 'room',
+    price: 182,
+    previewImage: 'https://15.design.htmlacademy.pro/static/hotel/16.jpg',
+    city: {
+      name: 'Amsterdam',
+      location: {
+        latitude: 52.37454,
+        longitude: 4.897976,
+        zoom: 13,
+      },
+    },
+    location: {
+      latitude: 52.35754,
+      longitude: 4.9179759999999995,
+      zoom: 16,
+    },
+    isFavorite: false,
+    isPremium: false,
+    rating: 2.9,
+  },
+  {
+    id: '7fa6eda3-4193-44aa-8955-97beecacde1f',
+    title: 'Nice, cozy, warm big bed apartment',
+    type: 'house',
+    price: 217,
+    previewImage: 'https://15.design.htmlacademy.pro/static/hotel/13.jpg',
+    city: {
+      name: 'Amsterdam',
+      location: {
+        latitude: 52.37454,
+        longitude: 4.897976,
+        zoom: 13,
+      },
+    },
+    location: {
+      latitude: 52.37854,
+      longitude: 4.894976,
+      zoom: 16,
+    },
+    isFavorite: false,
+    isPremium: true,
+    rating: 4.6,
+  },
+  {
+    id: '9a03d23c-d51a-4120-bfcb-df7fa1276ca1',
+    title: 'The Joshua Tree House',
+    type: 'room',
+    price: 267,
+    previewImage: 'https://15.design.htmlacademy.pro/static/hotel/17.jpg',
+    city: {
+      name: 'Cologne',
+      location: {
+        latitude: 50.938361,
+        longitude: 6.959974,
+        zoom: 13,
+      },
+    },
+    location: {
+      latitude: 50.945361,
+      longitude: 6.962974,
+      zoom: 16,
+    },
+    isFavorite: false,
+    isPremium: true,
+    rating: 3.7,
+  },
+  {
+    id: '20a98019-c99f-4016-888c-5ab101799038',
+    title: 'Canal View Prinsengracht',
+    type: 'house',
+    price: 858,
+    previewImage: 'https://15.design.htmlacademy.pro/static/hotel/19.jpg',
+    city: {
+      name: 'Cologne',
+      location: {
+        latitude: 50.938361,
+        longitude: 6.959974,
+        zoom: 13,
+      },
+    },
+    location: {
+      latitude: 50.917361,
+      longitude: 6.977974,
+      zoom: 16,
+    },
+    isFavorite: false,
+    isPremium: false,
+    rating: 3.8,
+  },
+  {
+    id: '9bce1e19-567a-432d-8571-9994832279b1',
+    title: 'Beautiful & luxurious apartment at great location',
+    type: 'house',
+    price: 863,
+    previewImage: 'https://15.design.htmlacademy.pro/static/hotel/9.jpg',
+    city: {
+      name: 'Paris',
+      location: {
+        latitude: 48.85661,
+        longitude: 2.351499,
+        zoom: 13,
+      },
+    },
+    location: {
+      latitude: 48.87961000000001,
+      longitude: 2.353499,
+      zoom: 16,
+    },
+    isFavorite: false,
+    isPremium: true,
+    rating: 2.2,
+  },
+];
