+++ src/pages/MainPage/MainPage.tsx
@@ -1,10 +1,11 @@
-import { useParams } from 'react-router-dom';
-import Navigation from '../../components/navigation/Navigation';
-import { mockOffersList } from '../../mockData';
 import { useEffect, useState } from 'react';
+import { useParams } from 'react-router-dom';
+
+import Navigation from '../../components/navigation/Navigation';
 import { TOffer } from '../../types';
 import CityPlaces from '../../components/cityPlaces/CityPlaces';
 import NoPlaces from '../../components/noPlaces/NoPlaces';
+import { mockOffersList } from '../../mockData/offers';
 
 function MainPage() {
   const { city } = useParams<{ city: string }>();
@@ -12,18 +13,14 @@ function MainPage() {
   const [offers, setOffers] = useState<TOffer[]>([]);
 
   useEffect(() => {
-    const offersByCity = mockOffersList.filter(
-      (offer) => offer.city.name.toLowerCase() === city,
-    );
+    const offersByCity = mockOffersList.filter((offer) => offer.city.name.toLowerCase() === city);
 
     setOffers(offersByCity);
   }, [city]);
 
   return (
     <main
-      className={`page__main page__main--index ${
-        !offers.length ? 'page__main--index-empty' : ''
-      }`}
+      className={`page__main page__main--index ${!offers.length ? 'page__main--index-empty' : ''}`}
     >
       <h1 className='visually-hidden'>Cities</h1>
       <Navigation />
